+++ src/app/contact/page.jsx
@@ -1,63 +1,100 @@
-import styles from "./page.module.css";
+/* eslint-disable jsx-a11y/alt-text */
+"use client";
 import Image from "next/image";
+import styles from "./page.module.css";
+import { useState } from "react";
 
-export default function Contact() {
+export default function ContactPage() {
+	const defaultState = {
+		name: "",
+		email: "",
+		subject: "Select...",
+		message: "",
+	};
+
+	const [formData, setFormData] = useState(defaultState);
+
+	const handleChange = (e) => {
+		const { name, value } = e.target;
+		setFormData({ ...formData, [name]: value });
+	};
+
+	const handleSubmit = (e) => {
+		e.preventDefault();
+		// Add submit logic here
+		console.log(formData);
+		setFormData(defaultState);
+	};
+
 	return (
-		<div className={styles.backgroundContainer}>
-			<h1 className={styles.mainTitle}>Contact Us</h1>
-			<main className={styles.mainPage}>
-				<p className={styles.contactText}>
-					Birmingham Business is an editorially-led magazine and we are keen to
-					hear from you, though the inclusion of articles or comment columns
-					cannot be guaranteed and is always at the Editor’s discretion.
-				</p>
-
-				<form className={styles.contactForm}>
-					<label for="name">
-						Name
-						<input className={styles.input} type="text" id="name" name="name" />
-					</label>
-					<label for="email">
-						Email{" "}
-						<input
-							className={styles.input}
-							type="email"
-							id="email"
-							name="email"
-						></input>
-					</label>
-					<label for="subject">
-						Subject{" "}
-						<select className={styles.input} name="select" id="issue">
-							<option value="dropdown">Select...</option>
-							<option value="editor">Editor</option>
-							<option value="webContent">Website Content</option>
-							<option value="advert">Advertising</option>
-							<option value="general">General Enquiries</option>
-							<option value="other">Other</option>
-						</select>
-					</label>
-					<label from="message">
-						Message{" "}
-						<textarea
-							className={styles.input}
-							name="message"
-							rows="10"
-							cols="50"
-						></textarea>
-					</label>
-					<button className={styles.submitButton}>Submit</button>
-					{/* <input className={styles.submitButton} type="submit" value="Submit" /> */}
-				</form>
-				<div className={styles.imageContainer}>
-					<Image
-						src={"/birmingham_background.svg"}
-						layout="fill"
-						className={styles.contactImage}
-						alt="Birmingham skyline background image"
+		<div className={styles.contactForm}>
+			<h1 className={styles.pageHeading}>Contact Us</h1>
+			<p>
+				Birmingham Business is an editorially-led magazine and we are keen to
+				hear from you, though the inclusion of articles or comment columns
+				cannot be guaranteed and is always at the Editor’s discretion.
+			</p>
+			<form onSubmit={handleSubmit} className={styles.formElement}>
+				<label className={styles.labelField} htmlFor="name">
+					Name
+					<input
+						onChange={handleChange}
+						className={styles.inputField}
+						type="text"
+						id="name"
+						name="name"
 					/>
+				</label>
+				<label className={styles.labelField} htmlFor="email">
+					Email
+					<input
+						onChange={handleChange}
+						className={styles.inputField}
+						type="email"
+						id="email"
+						name="email"
+					></input>
+				</label>
+				<label className={styles.labelField} htmlFor="subject">
+					Subject
+					<select
+						className={styles.inputField}
+						name="subject"
+						id="subject"
+						onChange={handleChange}
+					>
+						<option value="dropdown">Select...</option>
+						<option value="editor">Editor</option>
+						<option value="webContent">Website Content</option>
+						<option value="advert">Advertising</option>
+						<option value="general">General Enquiries</option>
+						<option value="other">Other</option>
+					</select>
+				</label>
+				<label className={styles.labelField} from="message">
+					Message
+					<textarea
+						onChange={handleChange}
+						className={styles.inputField}
+						name="message"
+						rows="10"
+						cols="40"
+					></textarea>
+				</label>
+				<div className={styles.buttonContainer}>
+					<button className={styles.button}>Send</button>
 				</div>
-			</main>
+			</form>
+			<div className={styles.imageContainer}>
+				<Image
+					src={"/birmingham_background.svg"}
+					width={100}
+					height={100}
+					layout="fixed"
+					className={styles.contactImage}
+					alt="Birmingham skyline background image"
+				/>
+			</div>
 		</div>
 	);
 }
